requirements.txt

reflex==0.8.9
mistralai
python-dotenv
fastapi
pymupdf
weasyprint
markdown2